LayoutTests/inspector/canvas/create-context-webgl-expected.txt

Test that CanvasManager tracks creation and destruction of WebGL canvases.


== Running test suite: Canvas.CreateContextWebGL
-- Running test case: Canvas.CreateContextWebGL.NoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: Canvas.CreateContextWebGL.Attached
PASS: Canvas context should be WebGL.
  0: getContext - [native code]
  1: createAttachedCanvas - inspector/canvas/resources/create-context-utilities.js:4:36
  2: Global Code - [program code]

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL.Detached
PASS: Canvas context should be WebGL.
  0: getContext - [native code]
  1: createDetachedCanvas - inspector/canvas/resources/create-context-utilities.js:13:62
  2: Global Code - [program code]

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL.Offscreen
PASS: Canvas context should be WebGL (Offscreen).
  0: getContext - [native code]
  1: createOffscreenCanvas - inspector/canvas/resources/create-context-utilities.js:36:36
  2: Global Code - [program code]

PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL.CSSCanvas
Create CSS canvas from -webkit-canvas(css-canvas).
PASS: Canvas context should be WebGL.
  0: getCSSCanvasContext - [native code]
  1: createCSSCanvas - inspector/canvas/resources/create-context-utilities.js:22:47
  2: Global Code - [program code]

PASS: Canvas name should equal the identifier passed to -webkit-canvas.
